feat: add AppSider
